backend/courses/requirements.txt
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
boto3==1.40.64
botocore==1.40.64
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
clickhouse-connect==0.9.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-etc==1.4.0
django-filter==25.2
django-hitcount==1.3.5
django-minio-storage==0.5.8
django-prometheus==2.4.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jmespath==1.0.1
lz4==4.4.4
minio==7.2.18
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
prometheus_client==0.23.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
s3transfer==0.14.0
sentry-sdk==2.43.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zope.interface==8.0.1
zstandard==0.25.0
